Refactor export tests into a new test_exports.py file
